Move registryMirrorsFlags to dedicated file
